Sensor AS7262 streaming/requirements.txt
PyQt5==5.*
pyserial==3.4
matplotlib
scipy
numpy
